stdlib: factor class path loading and native/1 out of StandardLibraryModuleLoader
